server/requirements.txt
flask
python-dotenv
flask-sqlalchemy
psycopg2-binary
bcrypt
pandas
shortuuid